Mnist dataset classification using CNN.
The MNIST database is a large database of handwritten digits that is commonly used for training various image processing systems. The database is also widely used for training and testing in the field of machine learning. Using convolutional neural networks, I am able to achieve 95% accuracy on the test set.